read the description
Whole front-end is working pretty well, also we got 200 status at process_payment post request. The only thing that seems to be not ok is the payment itself, the process_payment.php or shit like that. I need to watch that jerk doing that part because my frontend is working pretty well and retrieving the payment token, so i need to fix the back-end part. Hope the stupid helps, but i'm doing good so far.